+++ tests/src/test/java/com/jetbrains/youtrack/db/auto/DbImportExportTest.java
@@ -15,16 +15,16 @@
  */
 package com.jetbrains.youtrack.db.auto;
 
+import com.google.common.collect.ImmutableList;
 import com.jetbrains.youtrack.db.api.DatabaseType;
 import com.jetbrains.youtrack.db.api.YouTrackDB;
 import com.jetbrains.youtrack.db.api.YourTracks;
 import com.jetbrains.youtrack.db.api.config.GlobalConfiguration;
 import com.jetbrains.youtrack.db.api.config.YouTrackDBConfig;
+import com.jetbrains.youtrack.db.api.record.Entity;
 import com.jetbrains.youtrack.db.api.record.RID;
-import com.jetbrains.youtrack.db.api.record.RecordHook;
 import com.jetbrains.youtrack.db.api.schema.PropertyType;
 import com.jetbrains.youtrack.db.api.schema.Schema;
-import com.jetbrains.youtrack.db.api.schema.SchemaClass;
 import com.jetbrains.youtrack.db.internal.common.io.FileUtils;
 import com.jetbrains.youtrack.db.internal.common.log.LogManager;
 import com.jetbrains.youtrack.db.internal.core.command.CommandOutputListener;
@@ -34,14 +34,16 @@ import com.jetbrains.youtrack.db.internal.core.db.YouTrackDBImpl;
 import com.jetbrains.youtrack.db.internal.core.db.tool.DatabaseCompare;
 import com.jetbrains.youtrack.db.internal.core.db.tool.DatabaseExport;
 import com.jetbrains.youtrack.db.internal.core.db.tool.DatabaseImport;
-import com.jetbrains.youtrack.db.internal.core.id.RecordId;
-import com.jetbrains.youtrack.db.internal.core.record.impl.EntityImpl;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
+import java.util.Map.Entry;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 import org.testng.Assert;
+import org.testng.annotations.Ignore;
+import org.testng.annotations.Optional;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
@@ -57,46 +59,55 @@ public class DbImportExportTest extends BaseDBTest implements CommandOutputListe
   private boolean dumpMode = false;
 
   @Parameters(value = {"remote", "testPath"})
-  public DbImportExportTest(boolean remote, String testPath) {
-    super(remote);
+  public DbImportExportTest(@Optional Boolean remote, String testPath) {
+    super(remote != null && remote);
     this.testPath = testPath;
     this.exportFilePath = System.getProperty("exportFilePath", EXPORT_FILE_PATH);
   }
 
   @Test
   public void testDbExport() throws IOException {
-    // ADD A CUSTOM TO THE CLASS
-    database.command("alter class V custom onBeforeCreate=onBeforeCreateItem").close();
+    if (remoteDB) {
+      return;
+    }
 
-    final DatabaseExport export =
-        new DatabaseExport(database, testPath + "/" + exportFilePath, this);
+    // ADD A CUSTOM TO THE CLASS
+    session.execute("alter class V custom onBeforeCreate=onBeforeCreateItem").close();
+
+    final var export =
+        new DatabaseExport(session, testPath + "/" + exportFilePath, this);
     export.exportDatabase();
     export.close();
   }
 
   @Test(dependsOnMethods = "testDbExport")
   public void testDbImport() throws IOException {
-    final File importDir = new File(testPath + "/" + IMPORT_DB_PATH);
+    if (remoteDB) {
+      return;
+    }
+
+    final var importDir = new File(testPath + "/" + IMPORT_DB_PATH);
     if (importDir.exists()) {
-      for (final File f : importDir.listFiles()) {
+      for (final var f : importDir.listFiles()) {
         f.delete();
       }
     } else {
       importDir.mkdir();
     }
 
-    try (YouTrackDB youTrackDBImport =
+    try (var youTrackDBImport =
         YourTracks.embedded(
             testPath + File.separator + IMPORT_DB_PATH, YouTrackDBConfig.defaultConfig())) {
       youTrackDBImport.createIfNotExists(
-          IMPORT_DB_NAME, DatabaseType.PLOCAL, "admin", "admin", "admin");
+          IMPORT_DB_NAME, DatabaseType.DISK, "admin", "admin", "admin");
       try (var importDB = youTrackDBImport.open(IMPORT_DB_NAME, "admin", "admin")) {
-        final DatabaseImport dbImport =
+        final var dbImport =
             new DatabaseImport(
                 (DatabaseSessionInternal) importDB, testPath + "/" + exportFilePath, this);
         // UNREGISTER ALL THE HOOKS
-        for (final RecordHook hook : new ArrayList<>(database.getHooks().keySet())) {
-          database.unregisterHook(hook);
+        for (final var hook : new ArrayList<>(
+            ((DatabaseSessionInternal) importDB).getHooks())) {
+          session.unregisterHook(hook);
         }
         dbImport.setDeleteRIDMapping(false);
         dbImport.importDatabase();
@@ -108,18 +119,14 @@ public class DbImportExportTest extends BaseDBTest implements CommandOutputListe
   @Test(dependsOnMethods = "testDbImport")
   public void testCompareDatabases() throws IOException {
     if (remoteDB) {
-      final String env = getTestEnv();
-      if (env == null || env.equals("dev")) {
-        return;
-      }
-      // EXECUTES ONLY IF NOT REMOTE ON CI/RELEASE TEST ENV
+      return;
     }
-    try (YouTrackDB youTrackDBImport =
+    try (var youTrackDBImport =
         YourTracks.embedded(
             testPath + File.separator + IMPORT_DB_PATH, YouTrackDBConfig.defaultConfig())) {
       try (var importDB = youTrackDBImport.open(IMPORT_DB_NAME, "admin", "admin")) {
-        final DatabaseCompare databaseCompare =
-            new DatabaseCompare(database, (DatabaseSessionInternal) importDB, this);
+        final var databaseCompare =
+            new DatabaseCompare(session, (DatabaseSessionInternal) importDB, this);
         databaseCompare.setCompareEntriesForAutomaticIndexes(true);
         databaseCompare.setCompareIndexMetadata(true);
         Assert.assertTrue(databaseCompare.compare());
@@ -128,16 +135,16 @@ public class DbImportExportTest extends BaseDBTest implements CommandOutputListe
   }
 
   @Test
-  public void embeddedListMigration() throws Exception {
+  public void testLinksMigration() throws Exception {
     if (remoteDB) {
       return;
     }
 
-    final File localTesPath = new File(testPath + "/target", "embeddedListMigration");
+    final var localTesPath = new File(testPath + "/target", "embeddedListMigration");
     FileUtils.deleteRecursively(localTesPath);
     Assert.assertTrue(localTesPath.mkdirs());
 
-    final File exportPath = new File(localTesPath, "export.json.gz");
+    final var exportPath = new File(localTesPath, "export.json.gz");
 
     final YouTrackDBConfig config =
         new YouTrackDBConfigBuilderImpl()
@@ -146,79 +153,148 @@ public class DbImportExportTest extends BaseDBTest implements CommandOutputListe
     try (final YouTrackDB youTrackDB = new YouTrackDBImpl(
         "embedded:" + localTesPath.getPath(),
         config)) {
-      youTrackDB.create("original", DatabaseType.PLOCAL);
+      youTrackDB.create("original", DatabaseType.DISK);
 
-      try (final DatabaseSessionInternal session = (DatabaseSessionInternal) youTrackDB.open(
+      final var childDocCount = 50;
+
+      try (final var session = (DatabaseSessionInternal) youTrackDB.open(
           "original", "admin", "admin")) {
         final Schema schema = session.getMetadata().getSchema();
 
-        final SchemaClass rootCls = schema.createClass("RootClass");
-        rootCls.createProperty(session, "embeddedList", PropertyType.EMBEDDEDLIST);
+        final var rootCls = schema.createClass("RootClass");
+        rootCls.createProperty("no", PropertyType.INTEGER);
+        rootCls.createProperty("circular_link", PropertyType.LINK);
+        rootCls.createProperty("linkList", PropertyType.LINKLIST);
+        rootCls.createProperty("linkSet", PropertyType.LINKSET);
+        rootCls.createProperty("linkMap", PropertyType.LINKMAP);
 
-        final SchemaClass childCls = schema.createClass("ChildClass");
+        final var childCls = schema.createClass("ChildClass");
+        childCls.createProperty("no", PropertyType.INTEGER);
 
+        // creating and deleting some records to shift the next available IDs.
         final List<RID> ridsToDelete = new ArrayList<>();
-        for (int i = 0; i < 100; i++) {
-          session.begin();
-          final EntityImpl document = new EntityImpl(childCls);
-          document.save();
-          session.commit();
-
-          ridsToDelete.add(document.getIdentity());
+        for (var i = 0; i < 100; i++) {
+          ridsToDelete.add(
+              session.computeInTx(tx -> tx.newEntity(rootCls).getIdentity())
+          );
+          ridsToDelete.add(
+              session.computeInTx(tx -> tx.newEntity(childCls).getIdentity())
+          );
+        }
+        for (final var rid : ridsToDelete) {
+          session.executeInTx(tx -> tx.load(rid).delete());
         }
 
-        for (final RID rid : ridsToDelete) {
-          session.begin();
-          rid.getRecord().delete();
-          session.commit();
-        }
+        session.executeInTx(tx -> {
+          final var rootDoc1 = tx.newEntity(rootCls);
+          rootDoc1.setProperty("no", 1);
+          final var rootDoc2 = tx.newEntity(rootCls);
+          rootDoc2.setProperty("no", 2);
 
-        final EntityImpl rootDocument = new EntityImpl(rootCls);
-        final ArrayList<EntityImpl> documents = new ArrayList<>();
+          rootDoc1.setProperty("circular_link", rootDoc2.getIdentity());
+          rootDoc2.setProperty("circular_link", rootDoc1.getIdentity());
 
-        for (int i = 0; i < 10; i++) {
-          session.begin();
-          final EntityImpl embeddedDocument = new EntityImpl();
+          final var docList = rootDoc1.getOrCreateLinkList("linkList");
+          final var docSet = rootDoc1.getOrCreateLinkSet("linkSet");
+          final var docMap = rootDoc1.getOrCreateLinkMap("linkMap");
 
-          final EntityImpl doc = new EntityImpl(childCls);
-          doc.save();
-          session.commit();
+          for (var i = 0; i < childDocCount; i++) {
+            final var linkedDoc = tx.newEntity();
+            final var doc = tx.newEntity(childCls);
+            doc.setProperty("no", i);
+            linkedDoc.setProperty("link", doc.getIdentity());
+            linkedDoc.setProperty("no", i);
 
-          embeddedDocument.field("link", doc.getIdentity());
-          documents.add(embeddedDocument);
-        }
+            docList.add(linkedDoc);
 
-        session.begin();
-        rootDocument.field("embeddedList", documents);
-        rootDocument.save();
-        session.commit();
+            if (i % 2 == 0) {
+              docSet.add(linkedDoc);
+            }
 
-        final DatabaseExport databaseExport =
+            if (i % 3 == 0) {
+              docMap.put("" + i, linkedDoc);
+            }
+          }
+        });
+
+        final var databaseExport =
             new DatabaseExport(
                 session, exportPath.getPath(), System.out::println);
         databaseExport.exportDatabase();
       }
 
-      youTrackDB.create("imported", DatabaseType.PLOCAL);
-      try (final DatabaseSessionInternal session =
+      youTrackDB.create("imported", DatabaseType.DISK);
+      try (final var session =
           (DatabaseSessionInternal) youTrackDB.open("imported", "admin", "admin")) {
-        final DatabaseImport databaseImport =
+        final var databaseImport =
             new DatabaseImport(session, exportPath.getPath(), System.out::println);
         databaseImport.run();
 
-        final Iterator<EntityImpl> classIterator = session.browseClass("RootClass");
-        final EntityImpl rootDocument = classIterator.next();
+        session.executeInTx(tx -> {
+          final var rootDocs =
+              ImmutableList.copyOf(session.browseClass("RootClass"))
+                  .stream()
+                  .collect(Collectors.toMap(r -> r.getInt("no"), Function.identity()));
 
-        final List<EntityImpl> documents = rootDocument.field("embeddedList");
-        for (int i = 0; i < 10; i++) {
-          final EntityImpl embeddedDocument = documents.get(i);
+          Assert.assertEquals(rootDocs.size(), 2);
 
-          embeddedDocument.setLazyLoad(false);
-          final RecordId link = embeddedDocument.getProperty("link");
+          final var rootDoc1 = rootDocs.get(1);
+          final var rootDoc2 = rootDocs.get(2);
 
-          Assert.assertNotNull(link);
-          Assert.assertNotNull(link.getRecord());
-        }
+          Assert.assertNotNull(rootDoc1);
+          Assert.assertNotNull(rootDoc2);
+
+          Assert.assertEquals(rootDoc1.getLink("circular_link"), rootDoc2.getIdentity());
+          Assert.assertEquals(rootDoc2.getLink("circular_link"), rootDoc1.getIdentity());
+
+          final var docList = rootDoc1.getLinkList("linkList");
+
+          final var docListEntities = docList
+              .stream()
+              .map(tx::loadEntity)
+              .collect(Collectors.toMap(r -> r.getInt("no"), Function.identity()));
+
+          final var docSetEntities = rootDoc1.getLinkSet("linkSet")
+              .stream()
+              .map(tx::loadEntity)
+              .collect(Collectors.toMap(r -> r.getInt("no"), Function.identity()));
+
+          final var docMapEntities = rootDoc1.getLinkMap("linkMap")
+              .entrySet()
+              .stream()
+              .collect(Collectors.toMap(
+                  Entry::getKey,
+                  e -> tx.loadEntity(e.getValue().getIdentity())
+              ));
+
+          Assert.assertEquals(docListEntities.size(), 50);
+          Assert.assertEquals(docSetEntities.size(), Math.ceilDiv(childDocCount, 2));
+          Assert.assertEquals(docMapEntities.size(), Math.ceilDiv(childDocCount, 3));
+
+          for (var i = 0; i < childDocCount; i++) {
+
+            final var docId = docList.get(i).getIdentity();
+
+            final var docs = new ArrayList<Entity>();
+            docs.add(docListEntities.get(i));
+            if (i % 2 == 0) {
+              docs.add(docSetEntities.get(i));
+            }
+            if (i % 3 == 0) {
+              docs.add(docMapEntities.get("" + i));
+            }
+
+            for (var doc : docs) {
+              Assert.assertNotNull(doc);
+              Assert.assertEquals(doc.getIdentity(), docId);
+              Assert.assertEquals(doc.getInt("no"), i);
+            }
+
+            final var child = docs.getFirst().getEntity("link");
+            Assert.assertNotNull(child);
+            Assert.assertEquals(child.getInt("no"), i);
+          }
+        });
       }
     }
   }
